backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-multipart>=0.0.9
pypdf>=5.0.0
setuptools<82
numpy>=1.26.4
soundfile>=0.12.1
kokoro>=0.9.4
chatterbox-tts>=0.1.7
pytest>=8.0.0
